mac_requirements.txt
auto_mix_prep==0.2.0
Kivy==2.1.0
kivymd==1.0.2
plyer==2.0.0
ppadb-uiautomator==1.0.0
pyobjus==1.2.1